api/requirements.txt
# VulnScout Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
aiohttp==3.9.1
beautifulsoup4==4.12.3
lxml==5.1.0
dnspython==2.4.2
jinja2==3.1.3
aiofiles==23.2.1
websockets==12.0
aiosqlite==0.20.0